chore: :pencil2: fix typos and reorgnize the constant of the interaction
