+++ packages/s2-core/src/common/constant/basic.ts
@@ -34,3 +34,6 @@ export const PADDING_TOP = 0;
 export const PADDING_RIGHT = 1;
 export const PADDING_DOWN = 2;
 export const PADDING_LEFT = 3;
+
+export const MIN_CELL_WIDTH = 28;
+export const MIN_CELL_HEIGHT = 16;
