+++ packages/s2-core/src/interaction/row-col-resize.ts
@@ -4,17 +4,11 @@ import { ResizeInfo } from '../facet/header/interface';
 import { BaseInteraction } from './base';
 import { S2Event } from './events/types';
 import { SpreadSheet } from 'src/sheet-type';
-
-const MIN_CELL_WIDTH = 28;
-const MIN_CELL_HEIGHT = 16;
-
-export enum EventType {
-  ROW_W = 'spreadsheet:change-row-header-width',
-  COL_W = 'spreadsheet:change-column-header-width',
-  ROW_H = 'spreadsheet:change-row-header-height',
-  COL_H = 'spreadsheet:change-column-header-height',
-  TREE_W = 'spreadsheet:change-tree-width',
-}
+import {
+  ResizeEventType,
+  MIN_CELL_WIDTH,
+  MIN_CELL_HEIGHT,
+} from '@/common/constant';
 
 /**
  * Resize row&col width/height interaction
@@ -134,14 +128,14 @@ export class RowColResize extends BaseInteraction {
             children[0]?.attr('path')[0];
           const endPoint: ['M', number, number] = children[1]?.attr('path')[0];
 
-          let eventType: EventType;
+          let resizeEventType: ResizeEventType;
           let config: any;
           // todo，如何优化这段代码？
           if (info.type === 'col') {
             // eslint-disable-next-line default-case
             switch (info.affect) {
               case 'field':
-                eventType = EventType.ROW_W;
+                resizeEventType = ResizeEventType.ROW_W;
                 config = {
                   rowCfg: {
                     widthByField: {
@@ -151,7 +145,7 @@ export class RowColResize extends BaseInteraction {
                 };
                 break;
               case 'tree':
-                eventType = EventType.TREE_W;
+                resizeEventType = ResizeEventType.TREE_W;
                 config = {
                   rowCfg: {
                     treeRowsWidth: endPoint[1] - startPoint[1],
@@ -159,7 +153,7 @@ export class RowColResize extends BaseInteraction {
                 };
                 break;
               case 'cell':
-                eventType = EventType.COL_W;
+                resizeEventType = ResizeEventType.COL_W;
                 config = {
                   colCfg: {
                     widthByFieldValue: {
@@ -173,7 +167,7 @@ export class RowColResize extends BaseInteraction {
             // eslint-disable-next-line default-case
             switch (info.affect) {
               case 'field':
-                eventType = EventType.COL_H;
+                resizeEventType = ResizeEventType.COL_H;
                 config = {
                   colCfg: {
                     heightByField: {
@@ -184,7 +178,7 @@ export class RowColResize extends BaseInteraction {
                 break;
               case 'cell':
               case 'tree':
-                eventType = EventType.ROW_H;
+                resizeEventType = ResizeEventType.ROW_H;
                 config = {
                   cellCfg: {
                     height: endPoint[2] - startPoint[2],
@@ -193,7 +187,7 @@ export class RowColResize extends BaseInteraction {
                 break;
             }
           }
-          this.spreadsheet.emit(eventType, config);
+          this.spreadsheet.emit(resizeEventType, config);
           this.spreadsheet.setOptions(
             merge({}, this.spreadsheet.options, { style: config }),
           );
