+++ packages/s2-core/src/sheet-type/index.ts
@@ -57,8 +57,8 @@ import {
   BaseEvent,
   BrushSelection,
   RowColResize,
-  DataCellMutiSelection,
-  ColRowMutiSelection,
+  DataCellMultiSelection,
+  ColRowMultiSelection,
   DataCellClick,
   CornerTextClick,
   RowColumnClick,
@@ -619,11 +619,11 @@ export class SpreadSheet extends EE {
       );
       this.interactions.set(
         InteractionNames.DATA_CELL_MULTI_SELECTION_INTERACTION,
-        new DataCellMutiSelection(this),
+        new DataCellMultiSelection(this),
       );
       this.interactions.set(
         InteractionNames.COL_ROW_MULTI_SELECTION_INTERACTION,
-        new ColRowMutiSelection(this),
+        new ColRowMultiSelection(this),
       );
     }
   }
